Executes a maven goal

Maven goals may be executed with or without a pom.xml file. Some maven goals (e.g. compilation) require the use of a JDK.

execute_maven(
  goal,
  opts = c(),
  pom_path = NULL,
  quiet = .quietly(verbose),
  debug = .debug(verbose),
  verbose = c("normal", "debug", "quiet"),
  require_jdk = FALSE,
  settings = .settings_path(),
  ...
)

Arguments

goal

the goal of the mvn command ( can be multiple ) e.g. c("clean","compile")

opts

provided options in the form c("-Doption1=value2","-Doption2=value2")

pom_path

optional. the path to a pom.xml file for goals that operate on one

quiet

should output from maven be suppressed? (-q flag)

debug

should output from maven be verbose? (-X flag)

verbose

how much output from maven, one of "normal", "quiet", "debug"

require_jdk

does the goal you are executing require a JDK (e.g. compilation)

settings

the path to a settings.xml file controlling Maven. The default is a configuration with a local repository in the 'rmaven' cache directory (and not the Java maven repository).

...

named parameters are passed to maven as options in the form -Dname=value

Value

nothing, invisibly

Examples

# \donttest{
# This code can take quite a while to run as has to
# download a lot of plugins, especially on first run
execute_maven("help:system")
#> executing: /home/runner/.cache/rmaven/mvnw help:system -B -s '/tmp/RtmpSBBaiE/rmaven/settings.xml'
#> [INFO] Scanning for projects...
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-clean-plugin/2.5/maven-clean-plugin-2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-clean-plugin/2.5/maven-clean-plugin-2.5.pom (4 KB at 11.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/22/maven-plugins-22.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/22/maven-plugins-22.pom (13 KB at 397.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/21/maven-parent-21.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/21/maven-parent-21.pom (26 KB at 536.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/apache/10/apache-10.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/apache/10/apache-10.pom (15 KB at 437.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-clean-plugin/2.5/maven-clean-plugin-2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-clean-plugin/2.5/maven-clean-plugin-2.5.jar (25 KB at 801.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-install-plugin/2.4/maven-install-plugin-2.4.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-install-plugin/2.4/maven-install-plugin-2.4.pom (7 KB at 230.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/23/maven-plugins-23.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/23/maven-plugins-23.pom (9 KB at 345.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/22/maven-parent-22.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/22/maven-parent-22.pom (30 KB at 937.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/apache/11/apache-11.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/apache/11/apache-11.pom (15 KB at 602.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-install-plugin/2.4/maven-install-plugin-2.4.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-install-plugin/2.4/maven-install-plugin-2.4.jar (27 KB at 940.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-deploy-plugin/2.7/maven-deploy-plugin-2.7.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-deploy-plugin/2.7/maven-deploy-plugin-2.7.pom (6 KB at 210.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-deploy-plugin/2.7/maven-deploy-plugin-2.7.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-deploy-plugin/2.7/maven-deploy-plugin-2.7.jar (27 KB at 903.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-site-plugin/3.3/maven-site-plugin-3.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-site-plugin/3.3/maven-site-plugin-3.3.pom (21 KB at 781.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/24/maven-plugins-24.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/24/maven-plugins-24.pom (11 KB at 376.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-site-plugin/3.3/maven-site-plugin-3.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-site-plugin/3.3/maven-site-plugin-3.3.jar (122 KB at 2962.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-antrun-plugin/1.3/maven-antrun-plugin-1.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-antrun-plugin/1.3/maven-antrun-plugin-1.3.pom (5 KB at 184.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/12/maven-plugins-12.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/12/maven-plugins-12.pom (12 KB at 498.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/9/maven-parent-9.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/9/maven-parent-9.pom (33 KB at 1282.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-antrun-plugin/1.3/maven-antrun-plugin-1.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-antrun-plugin/1.3/maven-antrun-plugin-1.3.jar (24 KB at 885.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-assembly-plugin/2.2-beta-5/maven-assembly-plugin-2.2-beta-5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-assembly-plugin/2.2-beta-5/maven-assembly-plugin-2.2-beta-5.pom (15 KB at 616.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/16/maven-plugins-16.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/16/maven-plugins-16.pom (13 KB at 451.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-assembly-plugin/2.2-beta-5/maven-assembly-plugin-2.2-beta-5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-assembly-plugin/2.2-beta-5/maven-assembly-plugin-2.2-beta-5.jar (204 KB at 4079.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-dependency-plugin/2.8/maven-dependency-plugin-2.8.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-dependency-plugin/2.8/maven-dependency-plugin-2.8.pom (12 KB at 414.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-dependency-plugin/2.8/maven-dependency-plugin-2.8.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-dependency-plugin/2.8/maven-dependency-plugin-2.8.jar (150 KB at 4674.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-release-plugin/2.3.2/maven-release-plugin-2.3.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-release-plugin/2.3.2/maven-release-plugin-2.3.2.pom (10 KB at 378.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/release/maven-release/2.3.2/maven-release-2.3.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/release/maven-release/2.3.2/maven-release-2.3.2.pom (9 KB at 366.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-release-plugin/2.3.2/maven-release-plugin-2.3.2.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-release-plugin/2.3.2/maven-release-plugin-2.3.2.jar (44 KB at 1494.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-metadata.xml
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/mojo/maven-metadata.xml
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/mojo/maven-metadata.xml (21 KB at 671.5 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-metadata.xml (14 KB at 208.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-help-plugin/maven-metadata.xml
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-help-plugin/maven-metadata.xml (683 B at 20.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-help-plugin/3.3.0/maven-help-plugin-3.3.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-help-plugin/3.3.0/maven-help-plugin-3.3.0.pom (11 KB at 342.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/37/maven-plugins-37.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/37/maven-plugins-37.pom (10 KB at 333.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/37/maven-parent-37.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/37/maven-parent-37.pom (45 KB at 1536.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/apache/27/apache-27.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/apache/27/apache-27.pom (20 KB at 795.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-help-plugin/3.3.0/maven-help-plugin-3.3.0.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-help-plugin/3.3.0/maven-help-plugin-3.3.0.jar (83 KB at 2830.6 KB/sec)
#> [INFO]                                                                         
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Building Maven Stub Project (No POM) 1
#> [INFO] ------------------------------------------------------------------------
#> [INFO] 
#> [INFO] --- maven-help-plugin:3.3.0:system (default-cli) @ standalone-pom ---
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-generators/3.6.4/maven-plugin-tools-generators-3.6.4.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-generators/3.6.4/maven-plugin-tools-generators-3.6.4.pom (4 KB at 99.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools/3.6.4/maven-plugin-tools-3.6.4.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools/3.6.4/maven-plugin-tools-3.6.4.pom (15 KB at 434.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-api/3.6.4/maven-plugin-tools-api-3.6.4.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-api/3.6.4/maven-plugin-tools-api-3.6.4.pom (3 KB at 129.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-core/3.2.5/maven-core-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-core/3.2.5/maven-core-3.2.5.pom (8 KB at 207.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven/3.2.5/maven-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven/3.2.5/maven-3.2.5.pom (22 KB at 949.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/25/maven-parent-25.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/25/maven-parent-25.pom (37 KB at 1661.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/apache/15/apache-15.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/apache/15/apache-15.pom (15 KB at 330.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model/3.2.5/maven-model-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model/3.2.5/maven-model-3.2.5.pom (5 KB at 188.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-settings/3.2.5/maven-settings-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-settings/3.2.5/maven-settings-3.2.5.pom (3 KB at 92.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-settings-builder/3.2.5/maven-settings-builder-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-settings-builder/3.2.5/maven-settings-builder-3.2.5.pom (3 KB at 115.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interpolation/1.21/plexus-interpolation-1.21.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interpolation/1.21/plexus-interpolation-1.21.pom (2 KB at 57.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-repository-metadata/3.2.5/maven-repository-metadata-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-repository-metadata/3.2.5/maven-repository-metadata-3.2.5.pom (3 KB at 80.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-artifact/3.2.5/maven-artifact-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-artifact/3.2.5/maven-artifact-3.2.5.pom (3 KB at 91.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-plugin-api/3.2.5/maven-plugin-api-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-plugin-api/3.2.5/maven-plugin-api-3.2.5.pom (3 KB at 128.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.plexus/0.3.5/org.eclipse.sisu.plexus-0.3.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.plexus/0.3.5/org.eclipse.sisu.plexus-0.3.5.pom (5 KB at 190.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/sisu/sisu-plexus/0.3.5/sisu-plexus-0.3.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/sisu/sisu-plexus/0.3.5/sisu-plexus-0.3.5.pom (14 KB at 582.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/sonatype/oss/oss-parent/9/oss-parent-9.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/sonatype/oss/oss-parent/9/oss-parent-9.pom (7 KB at 229.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/javax/annotation/javax.annotation-api/1.2/javax.annotation-api-1.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/javax/annotation/javax.annotation-api/1.2/javax.annotation-api-1.2.pom (14 KB at 655.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/net/java/jvnet-parent/3/jvnet-parent-3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/net/java/jvnet-parent/3/jvnet-parent-3.pom (5 KB at 233.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/javax/enterprise/cdi-api/1.2/cdi-api-1.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/javax/enterprise/cdi-api/1.2/cdi-api-1.2.pom (7 KB at 291.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/jboss/weld/weld-parent/26/weld-parent-26.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/jboss/weld/weld-parent/26/weld-parent-26.pom (32 KB at 956.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.inject/0.3.5/org.eclipse.sisu.inject-0.3.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.inject/0.3.5/org.eclipse.sisu.inject-0.3.5.pom (3 KB at 98.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/sisu/sisu-inject/0.3.5/sisu-inject-0.3.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/sisu/sisu-inject/0.3.5/sisu-inject-0.3.5.pom (15 KB at 521.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-classworlds/2.5.2/plexus-classworlds-2.5.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-classworlds/2.5.2/plexus-classworlds-2.5.2.pom (8 KB at 310.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model-builder/3.2.5/maven-model-builder-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model-builder/3.2.5/maven-model-builder-3.2.5.pom (3 KB at 112.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-aether-provider/3.2.5/maven-aether-provider-3.2.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-aether-provider/3.2.5/maven-aether-provider-3.2.5.pom (5 KB at 166.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-api/1.0.0.v20140518/aether-api-1.0.0.v20140518.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-api/1.0.0.v20140518/aether-api-1.0.0.v20140518.pom (2 KB at 71.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether/1.0.0.v20140518/aether-1.0.0.v20140518.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether/1.0.0.v20140518/aether-1.0.0.v20140518.pom (30 KB at 1177.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-spi/1.0.0.v20140518/aether-spi-1.0.0.v20140518.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-spi/1.0.0.v20140518/aether-spi-1.0.0.v20140518.pom (3 KB at 80.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-util/1.0.0.v20140518/aether-util-1.0.0.v20140518.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-util/1.0.0.v20140518/aether-util-1.0.0.v20140518.pom (3 KB at 85.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-impl/1.0.0.v20140518/aether-impl-1.0.0.v20140518.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-impl/1.0.0.v20140518/aether-impl-1.0.0.v20140518.pom (4 KB at 135.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/sonatype/sisu/sisu-guice/3.2.3/sisu-guice-3.2.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/sonatype/sisu/sisu-guice/3.2.3/sisu-guice-3.2.3.pom (11 KB at 464.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/sonatype/sisu/inject/guice-parent/3.2.3/guice-parent-3.2.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/sonatype/sisu/inject/guice-parent/3.2.3/guice-parent-3.2.3.pom (14 KB at 230.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/sonatype/forge/forge-parent/38/forge-parent-38.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/sonatype/forge/forge-parent/38/forge-parent-38.pom (19 KB at 831.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava/16.0.1/guava-16.0.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava/16.0.1/guava-16.0.1.pom (6 KB at 313.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/16.0.1/guava-parent-16.0.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/16.0.1/guava-parent-16.0.1.pom (8 KB at 358.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-api/3.0/maven-reporting-api-3.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-api/3.0/maven-reporting-api-3.0.pom (3 KB at 105.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/shared/maven-shared-components/15/maven-shared-components-15.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/shared/maven-shared-components/15/maven-shared-components-15.pom (10 KB at 434.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/16/maven-parent-16.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/16/maven-parent-16.pom (23 KB at 710.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/apache/7/apache-7.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/apache/7/apache-7.pom (15 KB at 587.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/doxia/doxia-sink-api/1.0/doxia-sink-api-1.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/doxia/doxia-sink-api/1.0/doxia-sink-api-1.0.pom (2 KB at 64.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/doxia/doxia/1.0/doxia-1.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/doxia/doxia/1.0/doxia-1.0.pom (10 KB at 409.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/10/maven-parent-10.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-parent/10/maven-parent-10.pom (31 KB at 1287.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-commons/9.2/asm-commons-9.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-commons/9.2/asm-commons-9.2.pom (3 KB at 103.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-tree/9.2/asm-tree-9.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-tree/9.2/asm-tree-9.2.pom (3 KB at 126.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-analysis/9.2/asm-analysis-9.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-analysis/9.2/asm-analysis-9.2.pom (3 KB at 121.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/net/sf/jtidy/jtidy/r938/jtidy-r938.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/net/sf/jtidy/jtidy/r938/jtidy-r938.pom (10 KB at 428.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm/9.3/asm-9.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm/9.3/asm-9.3.pom (3 KB at 115.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-commons/9.3/asm-commons-9.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-commons/9.3/asm-commons-9.3.pom (3 KB at 103.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-tree/9.3/asm-tree-9.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-tree/9.3/asm-tree-9.3.pom (3 KB at 115.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-analysis/9.3/asm-analysis-9.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-analysis/9.3/asm-analysis-9.3.pom (3 KB at 111.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-api/3.1.1/maven-reporting-api-3.1.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-api/3.1.1/maven-reporting-api-3.1.1.pom (4 KB at 153.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-exec/1.6.0/maven-reporting-exec-1.6.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-exec/1.6.0/maven-reporting-exec-1.6.0.pom (14 KB at 645.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interactivity-api/1.1/plexus-interactivity-api-1.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interactivity-api/1.1/plexus-interactivity-api-1.1.pom (823 B at 36.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interactivity/1.1/plexus-interactivity-1.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interactivity/1.1/plexus-interactivity-1.1.pom (2 KB at 56.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-components/6.5/plexus-components-6.5.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-components/6.5/plexus-components-6.5.pom (3 KB at 138.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-utils/3.4.2/plexus-utils-3.4.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-utils/3.4.2/plexus-utils-3.4.2.pom (9 KB at 364.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/jdom/jdom2/2.0.6.1/jdom2-2.0.6.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/jdom/jdom2/2.0.6.1/jdom2-2.0.6.1.pom (5 KB at 204.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/thoughtworks/xstream/xstream/1.4.19/xstream-1.4.19.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/thoughtworks/xstream/xstream/1.4.19/xstream-1.4.19.pom (24 KB at 968.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/thoughtworks/xstream/xstream-parent/1.4.19/xstream-parent-1.4.19.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/thoughtworks/xstream/xstream-parent/1.4.19/xstream-parent-1.4.19.pom (43 KB at 1692.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/io/github/x-stream/mxparser/1.2.2/mxparser-1.2.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/io/github/x-stream/mxparser/1.2.2/mxparser-1.2.2.pom (24 KB at 1058.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/xmlpull/xmlpull/1.1.3.1/xmlpull-1.1.3.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/xmlpull/xmlpull/1.1.3.1/xmlpull-1.1.3.1.pom (386 B at 18.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-generators/3.6.4/maven-plugin-tools-generators-3.6.4.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-api/3.6.4/maven-plugin-tools-api-3.6.4.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model/3.2.5/maven-model-3.2.5.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/commons-collections/commons-collections/3.1/commons-collections-3.1.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/net/sf/jtidy/jtidy/r938/jtidy-r938.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-generators/3.6.4/maven-plugin-tools-generators-3.6.4.jar (49 KB at 1856.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm/9.3/asm-9.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model/3.2.5/maven-model-3.2.5.jar (158 KB at 2574.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-commons/9.3/asm-commons-9.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/plugin-tools/maven-plugin-tools-api/3.6.4/maven-plugin-tools-api-3.6.4.jar (24 KB at 360.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-tree/9.3/asm-tree-9.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm/9.3/asm-9.3.jar (120 KB at 1924.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-analysis/9.3/asm-analysis-9.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-tree/9.3/asm-tree-9.3.jar (52 KB at 577.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/commons-io/commons-io/2.6/commons-io-2.6.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-commons/9.3/asm-commons-9.3.jar (72 KB at 771.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-api/3.1.1/maven-reporting-api-3.1.1.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/ow2/asm/asm-analysis/9.3/asm-analysis-9.3.jar (34 KB at 356.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-exec/1.6.0/maven-reporting-exec-1.6.0.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/net/sf/jtidy/jtidy/r938/jtidy-r938.jar (245 KB at 2198.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-artifact/3.2.5/maven-artifact-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-api/3.1.1/maven-reporting-api-3.1.1.jar (11 KB at 90.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-core/3.2.5/maven-core-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/reporting/maven-reporting-exec/1.6.0/maven-reporting-exec-1.6.0.jar (31 KB at 258.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-repository-metadata/3.2.5/maven-repository-metadata-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/commons-io/commons-io/2.6/commons-io-2.6.jar (210 KB at 1565.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model-builder/3.2.5/maven-model-builder-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-repository-metadata/3.2.5/maven-repository-metadata-3.2.5.jar (26 KB at 171.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-aether-provider/3.2.5/maven-aether-provider-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-artifact/3.2.5/maven-artifact-3.2.5.jar (54 KB at 363.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-spi/1.0.0.v20140518/aether-spi-1.0.0.v20140518.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/commons-collections/commons-collections/3.1/commons-collections-3.1.jar (547 KB at 3310.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-impl/1.0.0.v20140518/aether-impl-1.0.0.v20140518.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-model-builder/3.2.5/maven-model-builder-3.2.5.jar (167 KB at 966.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-api/1.0.0.v20140518/aether-api-1.0.0.v20140518.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-aether-provider/3.2.5/maven-aether-provider-3.2.5.jar (65 KB at 350.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.plexus/0.3.5/org.eclipse.sisu.plexus-0.3.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-core/3.2.5/maven-core-3.2.5.jar (594 KB at 3159.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/javax/annotation/javax.annotation-api/1.2/javax.annotation-api-1.2.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-spi/1.0.0.v20140518/aether-spi-1.0.0.v20140518.jar (31 KB at 150.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/javax/enterprise/cdi-api/1.2/cdi-api-1.2.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/javax/annotation/javax.annotation-api/1.2/javax.annotation-api-1.2.jar (26 KB at 124.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.inject/0.3.5/org.eclipse.sisu.inject-0.3.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-impl/1.0.0.v20140518/aether-impl-1.0.0.v20140518.jar (168 KB at 769.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/sonatype/sisu/sisu-guice/3.2.3/sisu-guice-3.2.3-no_aop.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/javax/enterprise/cdi-api/1.2/cdi-api-1.2.jar (70 KB at 307.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava/16.0.1/guava-16.0.1.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.plexus/0.3.5/org.eclipse.sisu.plexus-0.3.5.jar (201 KB at 856.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interpolation/1.21/plexus-interpolation-1.21.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-api/1.0.0.v20140518/aether-api-1.0.0.v20140518.jar (134 KB at 564.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-classworlds/2.5.2/plexus-classworlds-2.5.2.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interpolation/1.21/plexus-interpolation-1.21.jar (61 KB at 234.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-plugin-api/3.2.5/maven-plugin-api-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-classworlds/2.5.2/plexus-classworlds-2.5.2.jar (52 KB at 194.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/maven/maven-settings/3.2.5/maven-settings-3.2.5.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/sisu/org.eclipse.sisu.inject/0.3.5/org.eclipse.sisu.inject-0.3.5.jar (371 KB at 1372.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-util/1.0.0.v20140518/aether-util-1.0.0.v20140518.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-plugin-api/3.2.5/maven-plugin-api-3.2.5.jar (46 KB at 158.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interactivity-api/1.1/plexus-interactivity-api-1.1.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/sonatype/sisu/sisu-guice/3.2.3/sisu-guice-3.2.3-no_aop.jar (389 KB at 1353.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-utils/3.4.2/plexus-utils-3.4.2.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/maven/maven-settings/3.2.5/maven-settings-3.2.5.jar (43 KB at 144.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/jdom/jdom2/2.0.6.1/jdom2-2.0.6.1.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/eclipse/aether/aether-util/1.0.0.v20140518/aether-util-1.0.0.v20140518.jar (143 KB at 477.9 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/thoughtworks/xstream/xstream/1.4.19/xstream-1.4.19.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-interactivity-api/1.1/plexus-interactivity-api-1.1.jar (10 KB at 29.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/io/github/x-stream/mxparser/1.2.2/mxparser-1.2.2.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/io/github/x-stream/mxparser/1.2.2/mxparser-1.2.2.jar (29 KB at 87.6 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/xmlpull/xmlpull/1.1.3.1/xmlpull-1.1.3.1.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/codehaus/plexus/plexus-utils/3.4.2/plexus-utils-3.4.2.jar (262 KB at 756.7 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/jdom/jdom2/2.0.6.1/jdom2-2.0.6.1.jar (321 KB at 919.9 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/xmlpull/xmlpull/1.1.3.1/xmlpull-1.1.3.1.jar (8 KB at 19.9 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/thoughtworks/xstream/xstream/1.4.19/xstream-1.4.19.jar (619 KB at 1585.4 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar (491 KB at 1198.3 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava/16.0.1/guava-16.0.1.jar (2176 KB at 5230.3 KB/sec)
#> [INFO] 
#> ===============================================================================
#> ========================= Platform Properties Details =========================
#> ===============================================================================
#> 
#> ===============================================================================
#> System Properties
#> ===============================================================================
#> 
#> awt.toolkit=sun.awt.X11.XToolkit
#> java.specification.version=11
#> sun.cpu.isalist=
#> sun.jnu.encoding=UTF-8
#> java.class.path=/home/runner/.cache/rmaven/.mvn/wrapper/maven-wrapper.jar
#> java.vm.vendor=Eclipse Adoptium
#> sun.arch.data.model=64
#> java.vendor.url=https://adoptium.net/
#> user.timezone=UTC
#> os.name=Linux
#> java.vm.specification.version=11
#> sun.java.launcher=SUN_STANDARD
#> sun.boot.library.path=/usr/lib/jvm/temurin-11-jdk-amd64/lib
#> sun.java.command=org.apache.maven.wrapper.MavenWrapperMain help:system -B -s /tmp/RtmpSBBaiE/rmaven/settings.xml
#> jdk.debug=release
#> maven.home=/home/runner/.m2/wrapper/dists/apache-maven-3.3.9-bin/6d5080f4/apache-maven-3.3.9
#> sun.cpu.endian=little
#> user.home=/home/runner
#> user.language=en
#> java.specification.vendor=Oracle Corporation
#> java.version.date=2022-08-12
#> java.home=/usr/lib/jvm/temurin-11-jdk-amd64
#> file.separator=/
#> java.vm.compressedOopsMode=32-bit
#> line.separator=
#> 
#> java.specification.name=Java Platform API Specification
#> java.vm.specification.vendor=Oracle Corporation
#> java.awt.graphicsenv=sun.awt.X11GraphicsEnvironment
#> sun.management.compiler=HotSpot 64-Bit Tiered Compilers
#> java.runtime.version=11.0.16.1+1
#> user.name=runner
#> path.separator=:
#> securerandom.source=file:/dev/./urandom
#> os.version=5.15.0-1019-azure
#> java.runtime.name=OpenJDK Runtime Environment
#> file.encoding=UTF-8
#> guice.disable.misplaced.annotation.check=true
#> java.vm.name=OpenJDK 64-Bit Server VM
#> java.vendor.version=Temurin-11.0.16.1+1
#> java.vendor.url.bug=https://github.com/adoptium/adoptium-support/issues
#> java.io.tmpdir=/tmp
#> java.version=11.0.16.1
#> user.dir=/home/runner/.cache/rmaven
#> os.arch=amd64
#> maven.multiModuleProjectDirectory=/home/runner/.cache/rmaven
#> java.vm.specification.name=Java Virtual Machine Specification
#> java.awt.printerjob=sun.print.PSPrinterJob
#> sun.os.patch.level=unknown
#> java.library.path=/usr/lib/jvm/temurin-11-jdk-amd64/lib/server:/usr/lib/jvm/temurin-11-jdk-amd64/lib:/usr/lib/jvm/temurin-11-jdk-amd64/../lib:/opt/R/4.2.1/lib/R/lib:/usr/local/lib:/usr/lib/x86_64-linux-gnu:/usr/lib/jvm/temurin-11-jdk-amd64/lib/server:/usr/java/packages/lib:/usr/lib64:/lib64:/lib:/usr/lib
#> java.vm.info=mixed mode
#> java.vendor=Eclipse Adoptium
#> java.vm.version=11.0.16.1+1
#> classworlds.conf=/home/runner/.m2/wrapper/dists/apache-maven-3.3.9-bin/6d5080f4/apache-maven-3.3.9/bin/m2.conf
#> sun.io.unicode.encoding=UnicodeLittle
#> java.class.version=55.0
#> 
#> ===============================================================================
#> Environment Variables
#> ===============================================================================
#> 
#> PATH=/home/linuxbrew/.linuxbrew/bin:/home/linuxbrew/.linuxbrew/sbin:/home/runner/.local/bin:/opt/pipx_bin:/home/runner/.cargo/bin:/home/runner/.config/composer/vendor/bin:/usr/local/.ghcup/bin:/home/runner/.dotnet/tools:/snap/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
#> GITHUB_REPOSITORY_OWNER=terminological
#> SWIFT_PATH=/usr/share/swift/usr/bin
#> SGX_AESM_ADDR=1
#> R_STRIP_SHARED_LIB=strip --strip-unneeded
#> CHROMEWEBDRIVER=/usr/local/share/chrome_driver
#> R_UNZIPCMD=/usr/bin/unzip
#> RUNNER_USER=runner
#> GITHUB_RUN_NUMBER=18
#> ANDROID_HOME=/usr/local/lib/android/sdk
#> GITHUB_RETENTION_DAYS=90
#> GITHUB_WORKSPACE=/home/runner/work/rmaven/rmaven
#> GITHUB_API_URL=https://api.github.com
#> LD_LIBRARY_PATH=/usr/lib/jvm/temurin-11-jdk-amd64/lib/server:/usr/lib/jvm/temurin-11-jdk-amd64/lib:/usr/lib/jvm/temurin-11-jdk-amd64/../lib:/opt/R/4.2.1/lib/R/lib:/usr/local/lib:/usr/lib/x86_64-linux-gnu:/usr/lib/jvm/temurin-11-jdk-amd64/lib/server
#> GITHUB_RUN_ATTEMPT=1
#> PWD=/home/runner/.cache/rmaven
#> GITHUB_JOB=pkgdown
#> GITHUB_ACTIONS=true
#> LANGUAGE=en
#> DOTNET_MULTILEVEL_LOOKUP=0
#> R_SHARE_DIR=/opt/R/4.2.1/lib/R/share
#> HOMEBREW_CLEANUP_PERIODIC_FULL_DAYS=3650
#> _R_CHECK_SYSTEM_CLOCK_=FALSE
#> PAGER=/usr/bin/pager
#> HOMEBREW_CELLAR=/home/linuxbrew/.linuxbrew/Cellar
#> ANDROID_NDK=/usr/local/lib/android/sdk/ndk/25.1.8937393
#> GOROOT_1_18_X64=/opt/hostedtoolcache/go/1.18.5/x64
#> GITHUB_ENV=/home/runner/work/_temp/_runner_file_commands/set_env_1bef9543-8075-4cc0-8498-faf98bb7ceef
#> GRAALVM_11_ROOT=/usr/local/graalvm/graalvm-ce-java11-22.2.0
#> GITHUB_EVENT_PATH=/home/runner/work/_temp/_github_workflow/event.json
#> SED=/usr/bin/sed
#> GITHUB_REF_TYPE=branch
#> R_PDFVIEWER=/usr/bin/xdg-open
#> RUNNER_TOOL_CACHE=/opt/hostedtoolcache
#> SHLVL=0
#> LEIN_HOME=/usr/local/lib/lein
#> R_LIBS_USER=/home/runner/work/_temp/Library
#> GITHUB_SERVER_URL=https://github.com
#> R_ZIPCMD=/usr/bin/zip
#> JAVA_HOME=/usr/lib/jvm/temurin-11-jdk-amd64
#> GITHUB_REF=refs/heads/main
#> RUNNER_NAME=GitHub Actions 7
#> R_STRIP_STATIC_LIB=strip --strip-debug
#> BOOTSTRAP_HASKELL_NONINTERACTIVE=1
#> HOMEBREW_PREFIX=/home/linuxbrew/.linuxbrew
#> GRADLE_HOME=/usr/share/gradle-7.5.1
#> ANT_HOME=/usr/share/ant
#> R_SYSTEM_ABI=linux,gcc,gxx,gfortran,gfortran
#> JAVA_HOME_17_X64=/usr/lib/jvm/temurin-17-jdk-amd64
#> GITHUB_STEP_SUMMARY=/home/runner/work/_temp/_runner_file_commands/step_summary_1bef9543-8075-4cc0-8498-faf98bb7ceef
#> GITHUB_REPOSITORY=terminological/rmaven
#> GITHUB_ACTION_REF=
#> HOMEBREW_REPOSITORY=/home/linuxbrew/.linuxbrew/Homebrew
#> R_BROWSER=xdg-open
#> AZURE_EXTENSION_DIR=/opt/az/azcliextensions
#> R_ARCH=
#> IN_PKGDOWN=true
#> CI=true
#> DOTNET_SKIP_FIRST_TIME_EXPERIENCE=1
#> LEIN_JAR=/usr/local/lib/lein/self-installs/leiningen-2.9.10-standalone.jar
#> MAVEN_PROJECTBASEDIR=/home/runner/.cache/rmaven
#> RUNNER_ARCH=X64
#> STATS_KEEPALIVE=false
#> GITHUB_SHA=7b72a093a0515cdfe9a2d25061ceb682f0f51e7f
#> ImageOS=ubuntu20
#> ImageVersion=20220905.1
#> ACCEPT_EULA=Y
#> GITHUB_PAT=ghs_qQgOw5gLytwmhFhBHpiDLkhhUoCdD81At0Sx
#> RUNNER_PERFLOG=/home/runner/perflog
#> R_INCLUDE_DIR=/opt/R/4.2.1/lib/R/include
#> PERFLOG_LOCATION_SETTING=RUNNER_PERFLOG
#> R_SESSION_TMPDIR=/tmp/RtmpSBBaiE
#> RUNNER_TRACKING_ID=github_f0a9e1ff-74a0-49a5-bca2-9f7fc16a6562
#> RUNNER_WORKSPACE=/home/runner/work/rmaven
#> R_GZIPCMD=/usr/bin/gzip
#> INVOCATION_ID=bc923a09b77e45e9b22026788e22f4da
#> TAR=/usr/bin/tar
#> TZ=UTC
#> GITHUB_REF_PROTECTED=false
#> EDGEWEBDRIVER=/usr/local/share/edge_driver
#> GITHUB_WORKFLOW=pkgdown
#> R_RD4PDF=times,inconsolata,hyper
#> CONDA=/usr/share/miniconda
#> SELENIUM_JAR_PATH=/usr/share/java/selenium-server.jar
#> R_PLATFORM=x86_64-pc-linux-gnu
#> GITHUB_BASE_REF=
#> PIPX_HOME=/opt/pipx
#> JOURNAL_STREAM=8:22305
#> R_DOC_DIR=/opt/R/4.2.1/lib/R/doc
#> LC_COLLATE=C
#> GITHUB_PATH=/home/runner/work/_temp/_runner_file_commands/add_path_1bef9543-8075-4cc0-8498-faf98bb7ceef
#> GITHUB_EVENT_NAME=push
#> GITHUB_TRIGGERING_ACTOR=robchallen
#> CHROME_BIN=/usr/bin/google-chrome
#> GOROOT_1_17_X64=/opt/hostedtoolcache/go/1.17.13/x64
#> AGENT_TOOLSDIRECTORY=/opt/hostedtoolcache
#> ANDROID_NDK_HOME=/usr/local/lib/android/sdk/ndk/25.1.8937393
#> ANDROID_NDK_ROOT=/usr/local/lib/android/sdk/ndk/25.1.8937393
#> VCPKG_INSTALLATION_ROOT=/usr/local/share/vcpkg
#> GITHUB_GRAPHQL_URL=https://api.github.com/graphql
#> R_PAPERSIZE=letter
#> RUNNER_OS=Linux
#> NOT_CRAN=true
#> DOTNET_NOLOGO=1
#> DEBIAN_FRONTEND=noninteractive
#> PKGCACHE_HTTP_VERSION=2
#> R_PRINTCMD=/usr/bin/lpr
#> GITHUB_ACTION=__run
#> POWERSHELL_DISTRIBUTION_CHANNEL=GitHub-Actions-ubuntu20
#> GITHUB_ACTOR=robchallen
#> GOROOT_1_19_X64=/opt/hostedtoolcache/go/1.19.0/x64
#> PIPX_BIN_DIR=/opt/pipx_bin
#> STATS_PFS=true
#> R_LIBS_SITE=/opt/R/4.2.1/lib/R/site-library
#> GECKOWEBDRIVER=/usr/local/share/gecko_driver
#> JAVA_HOME_8_X64=/usr/lib/jvm/temurin-8-jdk-amd64
#> MAKE=make
#> LANG=C.UTF-8
#> RSPM=https://packagemanager.rstudio.com/all/__linux__/focal/latest
#> R_BZIPCMD=/usr/bin/bzip2
#> MAVEN_CMD_LINE_ARGS= help:system -B -s /tmp/RtmpSBBaiE/rmaven/settings.xml
#> GITHUB_REF_NAME=main
#> NVM_DIR=/home/runner/.nvm
#> ANDROID_NDK_LATEST_HOME=/usr/local/lib/android/sdk/ndk/25.1.8937393
#> DEPLOYMENT_BASEPATH=/opt/runner
#> R_HOME=/opt/R/4.2.1/lib/R
#> HOMEBREW_NO_AUTO_UPDATE=1
#> GITHUB_ACTION_REPOSITORY=
#> ANDROID_SDK_ROOT=/usr/local/lib/android/sdk
#> GITHUB_RUN_ID=3062964307
#> USER=runner
#> XDG_CONFIG_HOME=/home/runner/.config
#> EDITOR=vi
#> R_TEXI2DVICMD=/usr/bin/texi2dvi
#> LN_S=ln -s
#> RUNNER_TEMP=/home/runner/work/_temp
#> GITHUB_HEAD_REF=
#> JAVA_HOME_11_X64=/usr/lib/jvm/temurin-11-jdk-amd64
#> XDG_RUNTIME_DIR=/run/user/1001
#> HOME=/home/runner
#> 
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD SUCCESS
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time: 4.735 s
#> [INFO] Finished at: 2022-09-15T19:01:57+00:00
#> [INFO] Final Memory: 13M/47M
#> [INFO] ------------------------------------------------------------------------
# }