Fetch an artifact from a repository into the local .m2 cache

Fetch an artifact from a repository into the local .m2 cache

fetch_artifact(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  ...,
  coordinates = NULL,
  artifact = NULL,
  repoUrl = getOption("rmaven.default_repos"),
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug")
)

Arguments

groupId

optional, the maven groupId,

artifactId

optional, the maven artifactId,

version

optional, the maven version,

...

other maven coordinates such as classifier or packaging

coordinates

optional, but if not supplied groupId and artifactId must be, coordinates as a coordinates object (see as.coordinates())

artifact

optional, coordinates as an artifact string groupId:artifactId:version[:packaging[:classifier]] string

repoUrl

the URLs of the repositories to check (defaults to maven central)

nocache

normally artifacts are only fetched if required, nocache forces fetching

verbose

how much output from maven, one of "normal", "quiet", "debug"

Value

the .m2 path of the artifact

Examples

fetch_artifact(artifact="io.github.terminological:r6-generator:main-SNAPSHOT:pom")
#> executing: /home/runner/.cache/rmaven/mvnw org.apache.maven.plugins:maven-dependency-plugin:3.3.0:get -Dartifact=io.github.terminological:r6-generator:main-SNAPSHOT:pom
#> Warning: running command ''/home/runner/.cache/rmaven/mvnw' org.apache.maven.plugins:maven-dependency-plugin:3.3.0:get -Dartifact=io.github.terminological:r6-generator:main-SNAPSHOT:pom' had status 1
#> [INFO] Scanning for projects...
#> [INFO]                                                                         
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Building Maven Stub Project (No POM) 1
#> [INFO] ------------------------------------------------------------------------
#> [INFO] 
#> [INFO] --- maven-dependency-plugin:3.3.0:get (default-cli) @ standalone-pom ---
#> [INFO] Resolving io.github.terminological:r6-generator:pom:main-SNAPSHOT with transitive dependencies
#> [WARNING] The POM for io.github.terminological:r6-generator:pom:main-SNAPSHOT is missing, no dependency information available
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD FAILURE
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time: 1.224 s
#> [INFO] Finished at: 2022-08-22T14:59:12+00:00
#> [INFO] Final Memory: 12M/44M
#> [INFO] ------------------------------------------------------------------------
#> [ERROR] Failed to execute goal org.apache.maven.plugins:maven-dependency-plugin:3.3.0:get (default-cli) on project standalone-pom: Couldn't download artifact: org.eclipse.aether.resolution.DependencyResolutionException: Could not find artifact io.github.terminological:r6-generator:pom:main-SNAPSHOT -> [Help 1]
#> [ERROR] 
#> [ERROR] To see the full stack trace of the errors, re-run Maven with the -e switch.
#> [ERROR] Re-run Maven using the -X switch to enable full debug logging.
#> [ERROR] 
#> [ERROR] For more information about the errors and possible solutions, please read the following articles:
#> [ERROR] [Help 1] http://cwiki.apache.org/confluence/display/MAVEN/MojoExecutionException
#> Error in fetch_artifact(artifact = "io.github.terminological:r6-generator:main-SNAPSHOT:pom"): did not sucessfully fetch artifact: /home/runner/.m2/repository/io/github/terminological/r6-generator/main-SNAPSHOT/r6-generator-main-SNAPSHOT.pom
fetch_artifact(coordinates = as.coordinates("org.junit.jupiter","junit-jupiter-api","5.9.0"))
#> /home/runner/.m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.0/junit-jupiter-api-5.9.0.jar