Resolve the classpath for an artifact

This calculates the dependencies for an artifact which may be specified either as a set of maven coordinates (in which case the artifact is downloaded) or as a path to a jar file containing a pom.xml (e.g. a compiled jar file, a compiled jar-with-dependencies, or a assembled src jar) The resulting file paths which will be in the maven local cache are checked on the filesystem.

resolve_dependencies(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  coordinates = NULL,
  path = NULL,
  include_self = NULL,
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug"),
  ...
)

Arguments

groupId

the maven groupId, optional (either groupId,artifactId and version must be specified, or coordinates)

artifactId

the maven artifactId, optional (either groupId,artifactId and version must be specified, or coordinates)

version

the maven version, optional (either groupId,artifactId and version must be specified, or coordinates)

coordinates

the maven coordinates, optional (either groupId,artifactId and version must be specified, or coordinates)

path

the path to the source directory, pom file or jar file

include_self

do you want include this path in the classpath. optional, if missing the path will be included if it is a regular jar, or a fat jar, otherwise not.

nocache

do not used cached version, by default we use a cached version of the classpath unless the pom.xml is newer that the cached classpath.

verbose

how much output from maven, one of "normal", "quiet", "debug"

...

passed on to as.coordinates()

Value

a character vector of the classpath jar files (including the current one if appropriate)

Examples

resolve_dependencies(groupId = "commons-io", artifactId = "commons-io", version="2.11.0")
#> executing: /home/runner/.cache/rmaven/mvnw dependency:build-classpath -Dmdep.outputFile=classpath.txt -DincludeScope=runtime
#> [1] "/home/runner/.m2/repository/commons-io/commons-io/2.11.0/commons-io-2.11.0.jar"

resolve_dependencies("org.junit.jupiter","junit-jupiter-api","5.9.0")
#> executing: /home/runner/.cache/rmaven/mvnw dependency:build-classpath -Dmdep.outputFile=classpath.txt -DincludeScope=runtime
#> [1] "/home/runner/.m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.0/junit-jupiter-api-5.9.0.jar"           
#> [2] "/home/runner/.m2/repository/org/opentest4j/opentest4j/1.2.0/opentest4j-1.2.0.jar"                            
#> [3] "/home/runner/.m2/repository/org/junit/platform/junit-platform-commons/1.9.0/junit-platform-commons-1.9.0.jar"
#> [4] "/home/runner/.m2/repository/org/apiguardian/apiguardian-api/1.1.2/apiguardian-api-1.1.2.jar"                 

resolve_dependencies(path=
  system.file("testdata/test-project-0.0.1-SNAPSHOT.jar",package="rmaven"))
#> executing: /home/runner/.cache/rmaven/mvnw dependency:build-classpath -Dmdep.outputFile=classpath.txt -DincludeScope=runtime
#> [1] "/home/runner/work/_temp/Library/rmaven/testdata/test-project-0.0.1-SNAPSHOT.jar"           
#> [2] "/home/runner/.m2/repository/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar"

resolve_dependencies(path=
  system.file("testdata/test-project-0.0.1-SNAPSHOT-src.jar",package="rmaven"))
#> executing: /home/runner/.cache/rmaven/mvnw dependency:build-classpath -Dmdep.outputFile=classpath.txt -DincludeScope=runtime
#> [1] "/home/runner/.m2/repository/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar"