Executes a maven goal

Maven goals are defined either as life-cycle goals (e.g. "clean", "compile") or as plugin goals (e.g. "help:system"). Some Maven goals may be executed without a pom.xml file, others require one. Some maven goals (e.g. compilation) require the use of a JDK.

execute_maven(
  goal,
  opts = c(),
  pom_path = NULL,
  quiet = .quietly(verbose),
  debug = .debug(verbose),
  verbose = c("normal", "debug", "quiet"),
  require_jdk = FALSE,
  settings = .settings_path(),
  ...
)

Arguments

goal

the goal of the mvn command ( can be multiple ) e.g. c("clean","compile")

opts

provided options in the form c("-Doption1=value2","-Doption2=value2")

pom_path

optional. the path to a pom.xml file for goals that need one.

quiet

should output from maven be suppressed? (-q flag)

debug

should output from maven be verbose? (-X flag)

verbose

how much output from maven, one of "normal", "quiet", "debug"

require_jdk

does the goal you are executing require a JDK (e.g. compilation does, fetching artifacts and calculating class path does not)

settings

the path to a settings.xml file controlling Maven. The default is a configuration with a local repository in the rmaven cache directory (and not the Java maven repository).

...

non-empty named parameters are passed to maven as options in the form -Dname=value

Value

nothing, invisibly

Examples

# \donttest{
# This code can take quite a while to run as has to
# download a lot of plugins, especially on first run on a clean system
execute_maven("help:system")
#> [INFO] Scanning for projects...
#> [INFO]                                                                         
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Building Maven Stub Project (No POM) 1
#> [INFO] ------------------------------------------------------------------------
#> [INFO] 
#> [INFO] --- maven-help-plugin:3.3.0:system (default-cli) @ standalone-pom ---
#> [INFO] 
#> ===============================================================================
#> ========================= Platform Properties Details =========================
#> ===============================================================================
#> 
#> ===============================================================================
#> System Properties
#> ===============================================================================
#> 
#> awt.toolkit=sun.awt.X11.XToolkit
#> java.specification.version=11
#> sun.cpu.isalist=
#> sun.jnu.encoding=UTF-8
#> java.class.path=/home/runner/.cache/rmaven/.mvn/wrapper/maven-wrapper.jar
#> java.vm.vendor=Eclipse Adoptium
#> sun.arch.data.model=64
#> java.vendor.url=https://adoptium.net/
#> user.timezone=UTC
#> os.name=Linux
#> java.vm.specification.version=11
#> sun.java.launcher=SUN_STANDARD
#> sun.boot.library.path=/usr/lib/jvm/temurin-11-jdk-amd64/lib
#> sun.java.command=org.apache.maven.wrapper.MavenWrapperMain help:system -B -s /tmp/RtmplOhmVg/rmaven/settings.xml -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn
#> jdk.debug=release
#> maven.home=/home/runner/.cache/rmaven/.m2/wrapper/dists/apache-maven-3.3.9-bin/6d5080f4/apache-maven-3.3.9
#> sun.cpu.endian=little
#> user.home=/home/runner
#> user.language=en
#> java.specification.vendor=Oracle Corporation
#> java.version.date=2024-01-16
#> java.home=/usr/lib/jvm/temurin-11-jdk-amd64
#> file.separator=/
#> java.vm.compressedOopsMode=Zero based
#> line.separator=
#> 
#> java.specification.name=Java Platform API Specification
#> java.vm.specification.vendor=Oracle Corporation
#> java.awt.graphicsenv=sun.awt.X11GraphicsEnvironment
#> sun.management.compiler=HotSpot 64-Bit Tiered Compilers
#> java.runtime.version=11.0.22+7
#> user.name=runner
#> org.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn
#> path.separator=:
#> securerandom.source=file:/dev/./urandom
#> os.version=6.5.0-1017-azure
#> java.runtime.name=OpenJDK Runtime Environment
#> file.encoding=UTF-8
#> guice.disable.misplaced.annotation.check=true
#> java.vm.name=OpenJDK 64-Bit Server VM
#> java.vendor.version=Temurin-11.0.22+7
#> java.vendor.url.bug=https://github.com/adoptium/adoptium-support/issues
#> java.io.tmpdir=/tmp
#> java.version=11.0.22
#> user.dir=/home/runner/.cache/rmaven
#> os.arch=amd64
#> maven.multiModuleProjectDirectory=/home/runner/.cache/rmaven
#> java.vm.specification.name=Java Virtual Machine Specification
#> java.awt.printerjob=sun.print.PSPrinterJob
#> sun.os.patch.level=unknown
#> java.library.path=/usr/lib/jvm/temurin-11-jdk-amd64/lib/server:/usr/lib/jvm/temurin-11-jdk-amd64/lib:/usr/lib/jvm/temurin-11-jdk-amd64/../lib:/opt/R/4.3.3/lib/R/lib:/usr/local/lib:/usr/lib/x86_64-linux-gnu:/usr/lib/jvm/temurin-11-jdk-amd64/lib/server:/usr/java/packages/lib:/usr/lib64:/lib64:/lib:/usr/lib
#> java.vm.info=mixed mode
#> java.vendor=Eclipse Adoptium
#> java.vm.version=11.0.22+7
#> java.specification.maintenance.version=2
#> classworlds.conf=/home/runner/.cache/rmaven/.m2/wrapper/dists/apache-maven-3.3.9-bin/6d5080f4/apache-maven-3.3.9/bin/m2.conf
#> sun.io.unicode.encoding=UnicodeLittle
#> java.class.version=55.0
#> 
#> ===============================================================================
#> Environment Variables
#> ===============================================================================
#> 
#> PATH=/opt/hostedtoolcache/pandoc/3.1.11/x64:/snap/bin:/home/runner/.local/bin:/opt/pipx_bin:/home/runner/.cargo/bin:/home/runner/.config/composer/vendor/bin:/usr/local/.ghcup/bin:/home/runner/.dotnet/tools:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
#> GITHUB_REPOSITORY_OWNER=terminological
#> SWIFT_PATH=/usr/share/swift/usr/bin
#> SGX_AESM_ADDR=1
#> R_STRIP_SHARED_LIB=strip --strip-unneeded
#> GITHUB_REPOSITORY_ID=525325654
#> CHROMEWEBDRIVER=/usr/local/share/chromedriver-linux64
#> R_UNZIPCMD=/usr/bin/unzip
#> RUNNER_USER=runner
#> GITHUB_RUN_NUMBER=42
#> ANDROID_HOME=/usr/local/lib/android/sdk
#> GITHUB_WORKSPACE=/home/runner/work/rmaven/rmaven
#> GITHUB_RETENTION_DAYS=90
#> GITHUB_API_URL=https://api.github.com
#> LD_LIBRARY_PATH=/usr/lib/jvm/temurin-11-jdk-amd64/lib/server:/usr/lib/jvm/temurin-11-jdk-amd64/lib:/usr/lib/jvm/temurin-11-jdk-amd64/../lib:/opt/R/4.3.3/lib/R/lib:/usr/local/lib:/usr/lib/x86_64-linux-gnu:/usr/lib/jvm/temurin-11-jdk-amd64/lib/server
#> GITHUB_RUN_ATTEMPT=1
#> PWD=/home/runner/.cache/rmaven
#> GITHUB_JOB=pkgdown
#> GITHUB_ACTIONS=true
#> STATS_V3PS=true
#> LANGUAGE=en
#> R_SHARE_DIR=/opt/R/4.3.3/lib/R/share
#> DOTNET_MULTILEVEL_LOOKUP=0
#> HOMEBREW_CLEANUP_PERIODIC_FULL_DAYS=3650
#> STATS_RDCL=true
#> _R_CHECK_SYSTEM_CLOCK_=FALSE
#> PAGER=/usr/bin/pager
#> RUNNER_ENVIRONMENT=github-hosted
#> ANDROID_NDK=/usr/local/lib/android/sdk/ndk/25.2.9519653
#> GITHUB_ENV=/home/runner/work/_temp/_runner_file_commands/set_env_9594da07-15aa-4c02-a861-67c3b9f7fd48
#> STATS_VMD=true
#> GITHUB_EVENT_PATH=/home/runner/work/_temp/_github_workflow/event.json
#> SED=/usr/bin/sed
#> STATS_D=true
#> GITHUB_ACTOR_ID=16591648
#> GITHUB_REF_TYPE=branch
#> R_PDFVIEWER=/usr/bin/xdg-open
#> RUNNER_TOOL_CACHE=/opt/hostedtoolcache
#> SHLVL=0
#> LEIN_HOME=/usr/local/lib/lein
#> R_LIBS_USER=/home/runner/work/_temp/Library
#> GITHUB_SERVER_URL=https://github.com
#> R_ZIPCMD=/usr/bin/zip
#> JAVA_HOME=/usr/lib/jvm/temurin-11-jdk-amd64
#> GITHUB_REF=refs/heads/main
#> RUNNER_NAME=GitHub Actions 6
#> R_STRIP_STATIC_LIB=strip --strip-debug
#> BOOTSTRAP_HASKELL_NONINTERACTIVE=1
#> STATS_TRP=true
#> GRADLE_HOME=/usr/share/gradle-8.7
#> ANT_HOME=/usr/share/ant
#> R_SYSTEM_ABI=linux,gcc,gxx,gfortran,gfortran
#> GITHUB_OUTPUT=/home/runner/work/_temp/_runner_file_commands/set_output_9594da07-15aa-4c02-a861-67c3b9f7fd48
#> JAVA_HOME_17_X64=/usr/lib/jvm/temurin-17-jdk-amd64
#> GITHUB_STEP_SUMMARY=/home/runner/work/_temp/_runner_file_commands/step_summary_9594da07-15aa-4c02-a861-67c3b9f7fd48
#> GITHUB_REPOSITORY=terminological/rmaven
#> GITHUB_ACTION_REF=
#> GOROOT_1_21_X64=/opt/hostedtoolcache/go/1.21.9/x64
#> R_BROWSER=xdg-open
#> AZURE_EXTENSION_DIR=/opt/az/azcliextensions
#> R_ARCH=
#> IN_PKGDOWN=true
#> CI=true
#> DOTNET_SKIP_FIRST_TIME_EXPERIENCE=1
#> LEIN_JAR=/usr/local/lib/lein/self-installs/leiningen-2.11.2-standalone.jar
#> MAVEN_PROJECTBASEDIR=/home/runner/.cache/rmaven
#> RUNNER_ARCH=X64
#> STATS_UE=true
#> GITHUB_SHA=acc644b3ff1bb1503496c57a4e8b97cbaf35f3e2
#> ImageOS=ubuntu22
#> ImageVersion=20240407.1.0
#> ACCEPT_EULA=Y
#> STATS_VMFE=true
#> GITHUB_PAT=ghs_rAa86LAfLcKUgztOqnT0S5RzCYGLeH21D86f
#> RUNNER_PERFLOG=/home/runner/perflog
#> R_INCLUDE_DIR=/opt/R/4.3.3/lib/R/include
#> PERFLOG_LOCATION_SETTING=RUNNER_PERFLOG
#> R_SESSION_TMPDIR=/tmp/RtmplOhmVg
#> RUNNER_TRACKING_ID=github_5e63f8dd-1229-4c42-822c-bde80ef8d708
#> RUNNER_WORKSPACE=/home/runner/work/rmaven
#> R_GZIPCMD=/usr/bin/gzip
#> INVOCATION_ID=d18d05d2082a4df580a54e289f4a12d4
#> TAR=/usr/bin/tar
#> TZ=UTC
#> GITHUB_REF_PROTECTED=false
#> EDGEWEBDRIVER=/usr/local/share/edge_driver
#> GITHUB_WORKFLOW=pkgdown
#> R_RD4PDF=times,inconsolata,hyper
#> CONDA=/usr/share/miniconda
#> SELENIUM_JAR_PATH=/usr/share/java/selenium-server.jar
#> R_PLATFORM=x86_64-pc-linux-gnu
#> GITHUB_BASE_REF=
#> PIPX_HOME=/opt/pipx
#> JOURNAL_STREAM=8:17264
#> GITHUB_WORKFLOW_SHA=acc644b3ff1bb1503496c57a4e8b97cbaf35f3e2
#> R_DOC_DIR=/opt/R/4.3.3/lib/R/doc
#> LC_COLLATE=C
#> GITHUB_PATH=/home/runner/work/_temp/_runner_file_commands/add_path_9594da07-15aa-4c02-a861-67c3b9f7fd48
#> ACTIONS_RUNNER_ACTION_ARCHIVE_CACHE=/opt/actionarchivecache
#> GITHUB_EVENT_NAME=push
#> GITHUB_TRIGGERING_ACTOR=robchallen
#> CHROME_BIN=/usr/bin/google-chrome
#> RENV_CONFIG_REPOS_OVERRIDE=https://packagemanager.posit.co/cran/__linux__/jammy/latest
#> AGENT_TOOLSDIRECTORY=/opt/hostedtoolcache
#> ANDROID_NDK_HOME=/usr/local/lib/android/sdk/ndk/25.2.9519653
#> ANDROID_NDK_ROOT=/usr/local/lib/android/sdk/ndk/25.2.9519653
#> VCPKG_INSTALLATION_ROOT=/usr/local/share/vcpkg
#> GITHUB_GRAPHQL_URL=https://api.github.com/graphql
#> R_PAPERSIZE=letter
#> JAVA_HOME_21_X64=/usr/lib/jvm/temurin-21-jdk-amd64
#> RUNNER_OS=Linux
#> NOT_CRAN=true
#> DOTNET_NOLOGO=1
#> DEBIAN_FRONTEND=noninteractive
#> PKGCACHE_HTTP_VERSION=2
#> GITHUB_WORKFLOW_REF=terminological/rmaven/.github/workflows/pkgdown.yaml@refs/heads/main
#> R_PRINTCMD=/usr/bin/lpr
#> GITHUB_ACTION=__run
#> POWERSHELL_DISTRIBUTION_CHANNEL=GitHub-Actions-ubuntu22
#> GITHUB_ACTOR=robchallen
#> SYSTEMD_EXEC_PID=598
#> PIPX_BIN_DIR=/opt/pipx_bin
#> R_LIBS_SITE=/opt/R/4.3.3/lib/R/site-library
#> MAVEN_USER_HOME=/home/runner/.cache/rmaven/.m2
#> GECKOWEBDRIVER=/usr/local/share/gecko_driver
#> GOROOT_1_22_X64=/opt/hostedtoolcache/go/1.22.2/x64
#> JAVA_HOME_8_X64=/usr/lib/jvm/temurin-8-jdk-amd64
#> MAKE=make
#> LANG=C.UTF-8
#> GITHUB_REPOSITORY_OWNER_ID=16591656
#> STATS_D_D=true
#> STATS_EXTP=https://provjobdsettingscdn.blob.core.windows.net/settings/provjobdsettings-0.5.172+1/provjobd.data
#> RSPM=https://packagemanager.posit.co/cran/__linux__/jammy/latest
#> GHCUP_INSTALL_BASE_PREFIX=/usr/local
#> R_BZIPCMD=/usr/bin/bzip2
#> STATS_EXT=true
#> GITHUB_STATE=/home/runner/work/_temp/_runner_file_commands/save_state_9594da07-15aa-4c02-a861-67c3b9f7fd48
#> MAVEN_CMD_LINE_ARGS= help:system -B -s /tmp/RtmplOhmVg/rmaven/settings.xml -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn
#> GITHUB_REF_NAME=main
#> NVM_DIR=/home/runner/.nvm
#> ANDROID_NDK_LATEST_HOME=/usr/local/lib/android/sdk/ndk/26.2.11394342
#> DEPLOYMENT_BASEPATH=/opt/runner
#> R_HOME=/opt/R/4.3.3/lib/R
#> HOMEBREW_NO_AUTO_UPDATE=1
#> GITHUB_ACTION_REPOSITORY=
#> ANDROID_SDK_ROOT=/usr/local/lib/android/sdk
#> GITHUB_RUN_ID=8618882658
#> GOROOT_1_20_X64=/opt/hostedtoolcache/go/1.20.14/x64
#> USER=runner
#> XDG_CONFIG_HOME=/home/runner/.config
#> EDITOR=vi
#> R_TEXI2DVICMD=/usr/bin/texi2dvi
#> LN_S=ln -s
#> RUNNER_TEMP=/home/runner/work/_temp
#> GITHUB_HEAD_REF=
#> JAVA_HOME_11_X64=/usr/lib/jvm/temurin-11-jdk-amd64
#> XDG_RUNTIME_DIR=/run/user/1001
#> HOME=/home/runner
#> R_LIB_FOR_PAK=/opt/R/4.3.3/lib/R/site-library
#> 
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD SUCCESS
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time: 2.992 s
#> [INFO] Finished at: 2024-04-09T16:11:10+00:00
#> [INFO] Final Memory: 13M/50M
#> [INFO] ------------------------------------------------------------------------
# }