Fetch an artifact from a repository into the local .m2 cache

Fetch an artifact from a repository into the local .m2 cache

fetch_artifact(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  ...,
  coordinates = NULL,
  artifact = NULL,
  repoUrl = getOption("rmaven.default_repos"),
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug")
)

Arguments

groupId

optional, the maven groupId,

artifactId

optional, the maven artifactId,

version

optional, the maven version,

...

other maven coordinates such as classifier or packaging

coordinates

optional, but if not supplied groupId and artifactId must be, coordinates as a coordinates object (see as.coordinates())

artifact

optional, coordinates as an artifact string groupId:artifactId:version[:packaging[:classifier]] string

repoUrl

the URLs of the repositories to check (defaults to maven central)

nocache

normally artifacts are only fetched if required, nocache forces fetching

verbose

how much output from maven, one of "normal", "quiet", "debug"

Value

the .m2 path of the artifact

Examples

fetch_artifact(artifact="com.google.guava:guava:31.1-jre")
#> executing: /home/runner/.cache/rmaven/mvnw org.apache.maven.plugins:maven-dependency-plugin:3.3.0:get -Dartifact=com.google.guava:guava:31.1-jre
#> [INFO] Scanning for projects...
#> [INFO]                                                                         
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Building Maven Stub Project (No POM) 1
#> [INFO] ------------------------------------------------------------------------
#> [INFO] 
#> [INFO] --- maven-dependency-plugin:3.3.0:get (default-cli) @ standalone-pom ---
#> [INFO] Resolving com.google.guava:guava:jar:31.1-jre with transitive dependencies
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.pom
#> 2/11 KB   3/11 KB   5/11 KB   6/11 KB   7/11 KB   9/11 KB   10/11 KB   11/11 KB   11/11 KB              Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.pom (11 KB at 37.5 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/31.1-jre/guava-parent-31.1-jre.pom
#> 2/15 KB    3/15 KB   5/15 KB   6/15 KB   7/15 KB   9/15 KB   10/15 KB   11/15 KB   13/15 KB   14/15 KB   15/15 KB              Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/31.1-jre/guava-parent-31.1-jre.pom (15 KB at 351.0 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.pom
#> 2/3 KB     3/3 KB            Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.pom (3 KB at 69.3 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/26.0-android/guava-parent-26.0-android.pom
#> 2/10 KB   3/10 KB   5/10 KB   6/10 KB   7/10 KB   9/10 KB   10/10 KB   10/10 KB              Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/26.0-android/guava-parent-26.0-android.pom (10 KB at 284.0 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.pom
#> 2/3 KB     3/3 KB            Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.pom (3 KB at 69.5 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.pom
#> 2/5 KB   3/5 KB   5/5 KB   5/5 KB            Downloaded: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.pom (5 KB at 113.1 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.pom
#> 2/3 KB   3/3 KB            Downloaded: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.pom (3 KB at 55.2 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.pom
#> 2/3 KB   3/3 KB            Downloaded: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.pom (3 KB at 60.4 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_parent/2.11.0/error_prone_parent-2.11.0.pom
#> 4/11 KB   8/11 KB   11/11 KB              Downloaded: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_parent/2.11.0/error_prone_parent-2.11.0.pom (11 KB at 268.8 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.pom
#> 3/3 KB              Downloaded: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.pom (3 KB at 74.9 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar
#> Downloading: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.jar
#> Downloading: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.jar
#> Downloading: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar
#> 4/2891 KB   8/2891 KB   12/2891 KB   16/2891 KB   20/2891 KB   24/2891 KB   28/2891 KB   32/2891 KB   36/2891 KB   40/2891 KB   44/2891 KB   48/2891 KB   52/2891 KB   56/2891 KB   60/2891 KB   64/2891 KB   68/2891 KB   72/2891 KB   76/2891 KB   80/2891 KB   84/2891 KB   88/2891 KB   92/2891 KB   96/2891 KB   100/2891 KB   104/2891 KB   108/2891 KB   112/2891 KB   116/2891 KB   120/2891 KB   124/2891 KB   128/2891 KB   132/2891 KB   136/2891 KB   140/2891 KB   144/2891 KB   148/2891 KB   152/2891 KB   156/2891 KB   160/2891 KB   164/2891 KB   168/2891 KB   172/2891 KB   176/2891 KB   180/2891 KB   184/2891 KB   188/2891 KB   192/2891 KB   196/2891 KB   200/2891 KB   204/2891 KB   208/2891 KB   212/2891 KB   2/204 KB   212/2891 KB   3/204 KB   212/2891 KB   5/204 KB   212/2891 KB   6/204 KB   212/2891 KB   7/204 KB   212/2891 KB   9/204 KB   212/2891 KB   10/204 KB   212/2891 KB   11/204 KB   212/2891 KB   13/204 KB   212/2891 KB   14/204 KB   212/2891 KB   15/204 KB   212/2891 KB   16/204 KB   212/2891 KB   16/204 KB   216/2891 KB   16/204 KB   220/2891 KB   16/204 KB   224/2891 KB   16/204 KB   228/2891 KB   16/204 KB   232/2891 KB   16/204 KB   236/2891 KB   16/204 KB   240/2891 KB   16/204 KB   244/2891 KB   16/204 KB   248/2891 KB   16/204 KB   252/2891 KB   16/204 KB   256/2891 KB   16/204 KB   260/2891 KB   16/204 KB   264/2891 KB   16/204 KB   268/2891 KB   16/204 KB   272/2891 KB   16/204 KB   276/2891 KB   16/204 KB   280/2891 KB   20/204 KB   284/2891 KB   24/204 KB   284/2891 KB   28/204 KB   284/2891 KB   32/204 KB   284/2891 KB   2/3 KB   32/204 KB   284/2891 KB   3/3 KB   32/204 KB   284/2891 KB   3/3 KB   32/204 KB   284/2891 KB   2/20 KB   3/3 KB   32/204 KB   284/2891 KB   3/20 KB   3/3 KB   32/204 KB   284/2891 KB   5/20 KB   3/3 KB   32/204 KB   284/2891 KB   6/20 KB   3/3 KB   32/204 KB   284/2891 KB   7/20 KB   3/3 KB   32/204 KB   284/2891 KB   9/20 KB   3/3 KB   32/204 KB   284/2891 KB   10/20 KB   3/3 KB   32/204 KB   284/2891 KB   11/20 KB   3/3 KB   32/204 KB   284/2891 KB   13/20 KB   3/3 KB   32/204 KB   284/2891 KB   14/20 KB   3/3 KB   32/204 KB   284/2891 KB   15/20 KB   3/3 KB   32/204 KB   284/2891 KB   16/20 KB   3/3 KB   32/204 KB   284/2891 KB   17/20 KB   3/3 KB   32/204 KB   284/2891 KB   19/20 KB   3/3 KB   32/204 KB   284/2891 KB   20/20 KB   16/204 KB   284/2891 KB                       3/3 KB   32/204 KB   288/2891 KB   20/20 KB   3/3 KB   32/204 KB   292/2891 KB   20/20 KB   3/3 KB   32/204 KB   296/2891 KB   20/20 KB   3/3 KB   36/204 KB   296/2891 KB   20/20 KB   3/3 KB   40/204 KB   296/2891 KB   20/20 KB   3/3 KB   44/204 KB   296/2891 KB   20/20 KB   3/3 KB   48/204 KB   296/2891 KB   20/20 KB   3/3 KB   48/204 KB   300/2891 KB   20/20 KB   3/3 KB   48/204 KB   304/2891 KB   20/20 KB   3/3 KB   48/204 KB   308/2891 KB   20/20 KB   3/3 KB   48/204 KB   312/2891 KB   20/20 KB   3/3 KB   48/204 KB   316/2891 KB   20/20 KB   3/3 KB   48/204 KB   320/2891 KB   20/20 KB   3/3 KB   52/204 KB   320/2891 KB   20/20 KB   3/3 KB   56/204 KB   320/2891 KB   20/20 KB   3/3 KB   60/204 KB   320/2891 KB   20/20 KB   3/3 KB   64/204 KB   320/2891 KB   20/20 KB   2/5 KB   3/3 KB   64/204 KB   320/2891 KB   20/20 KB   3/5 KB   3/3 KB   64/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   64/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   64/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   68/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   72/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   76/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   320/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   324/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   328/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   332/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   336/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   340/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   344/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   348/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   352/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   356/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   360/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   364/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   368/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   372/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   376/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   380/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   384/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   388/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   392/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   396/2891 KB   20/20 KB   5/5 KB   3/3 KB   80/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   84/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   88/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   92/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   96/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   100/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   104/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   108/204 KB   400/2891 KB   20/20 KB   5/5 KB   3/3 KB   112/204 KB   400/2891 KB   20/20 KB                                                           Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar (3 KB at 20.5 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.jar
#>                                                         Downloaded: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.jar (20 KB at 183.7 KB/sec)
#> Downloading: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.jar
#> 5/5 KB   116/204 KB   400/2891 KB                       5/5 KB   120/204 KB   400/2891 KB   5/5 KB   124/204 KB   400/2891 KB   5/5 KB   128/204 KB   400/2891 KB   5/5 KB   132/204 KB   400/2891 KB   5/5 KB   136/204 KB   400/2891 KB   5/5 KB   140/204 KB   400/2891 KB   5/5 KB   144/204 KB   400/2891 KB   5/5 KB   148/204 KB   400/2891 KB   5/5 KB   152/204 KB   400/2891 KB   5/5 KB   156/204 KB   400/2891 KB   5/5 KB   160/204 KB   400/2891 KB   5/5 KB   164/204 KB   400/2891 KB   5/5 KB   168/204 KB   400/2891 KB   5/5 KB   172/204 KB   400/2891 KB   5/5 KB   176/204 KB   400/2891 KB   5/5 KB   180/204 KB   400/2891 KB   5/5 KB   184/204 KB   404/2891 KB   5/5 KB   188/204 KB   404/2891 KB   5/5 KB   192/204 KB   404/2891 KB   5/5 KB   180/204 KB   404/2891 KB   5/5 KB   192/204 KB   408/2891 KB   5/5 KB   192/204 KB   412/2891 KB   5/5 KB   192/204 KB   416/2891 KB                                       Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar (5 KB at 39.6 KB/sec)
#> 192/204 KB   420/2891 KB            192/204 KB   424/2891 KB   196/204 KB   428/2891 KB   196/204 KB   432/2891 KB   196/204 KB   436/2891 KB   196/204 KB   440/2891 KB   196/204 KB   444/2891 KB   196/204 KB   448/2891 KB   196/204 KB   452/2891 KB   196/204 KB   456/2891 KB   196/204 KB   460/2891 KB   196/204 KB   464/2891 KB   196/204 KB   468/2891 KB   196/204 KB   424/2891 KB   200/204 KB   468/2891 KB   204/204 KB   468/2891 KB   204/204 KB   468/2891 KB   2/16 KB   204/204 KB   468/2891 KB   3/16 KB   204/204 KB   468/2891 KB   5/16 KB   204/204 KB   468/2891 KB   6/16 KB   204/204 KB   468/2891 KB   7/16 KB   204/204 KB   468/2891 KB   9/16 KB   204/204 KB   468/2891 KB   10/16 KB   204/204 KB   468/2891 KB   2/9 KB   10/16 KB   204/204 KB   468/2891 KB   3/9 KB   10/16 KB   204/204 KB   468/2891 KB   5/9 KB   10/16 KB   204/204 KB   468/2891 KB   6/9 KB   10/16 KB   204/204 KB   468/2891 KB   7/9 KB   10/16 KB   204/204 KB   468/2891 KB   9/9 KB   10/16 KB   204/204 KB   468/2891 KB   9/9 KB   10/16 KB   204/204 KB   468/2891 KB   9/9 KB   11/16 KB   204/204 KB   468/2891 KB   9/9 KB   13/16 KB   204/204 KB   468/2891 KB   9/9 KB   14/16 KB   204/204 KB   468/2891 KB   9/9 KB   15/16 KB   204/204 KB   468/2891 KB   9/9 KB   16/16 KB   204/204 KB   468/2891 KB   9/9 KB   16/16 KB   204/204 KB   468/2891 KB   9/9 KB   16/16 KB   204/204 KB   472/2891 KB   9/9 KB   16/16 KB   204/204 KB   476/2891 KB   9/9 KB   16/16 KB   204/204 KB   480/2891 KB   9/9 KB   16/16 KB   204/204 KB   484/2891 KB   9/9 KB   16/16 KB   204/204 KB   488/2891 KB   9/9 KB   16/16 KB   204/204 KB   492/2891 KB   9/9 KB   16/16 KB   204/204 KB   496/2891 KB   9/9 KB   16/16 KB   204/204 KB   500/2891 KB   9/9 KB   16/16 KB   204/204 KB   504/2891 KB   9/9 KB   16/16 KB   204/204 KB   508/2891 KB   9/9 KB   16/16 KB   204/204 KB   512/2891 KB   9/9 KB   16/16 KB   204/204 KB   516/2891 KB   9/9 KB   16/16 KB   204/204 KB   520/2891 KB   9/9 KB   16/16 KB   204/204 KB   524/2891 KB   9/9 KB   16/16 KB   204/204 KB   528/2891 KB   9/9 KB   16/16 KB   204/204 KB   532/2891 KB   9/9 KB   16/16 KB   204/204 KB   536/2891 KB   9/9 KB   16/16 KB   204/204 KB   540/2891 KB   9/9 KB   16/16 KB   204/204 KB   544/2891 KB   9/9 KB   16/16 KB   204/204 KB   548/2891 KB   9/9 KB   16/16 KB   204/204 KB   552/2891 KB   9/9 KB   16/16 KB   204/204 KB   556/2891 KB   9/9 KB   16/16 KB   204/204 KB   560/2891 KB   9/9 KB   16/16 KB   204/204 KB   564/2891 KB   9/9 KB   16/16 KB   204/204 KB   568/2891 KB   9/9 KB   16/16 KB   204/204 KB   571/2891 KB   9/9 KB   16/16 KB   204/204 KB   575/2891 KB   9/9 KB   16/16 KB   204/204 KB   579/2891 KB   9/9 KB   16/16 KB   204/204 KB   583/2891 KB   9/9 KB   16/16 KB   204/204 KB   587/2891 KB   9/9 KB   16/16 KB   204/204 KB   591/2891 KB   9/9 KB   16/16 KB   204/204 KB   595/2891 KB   9/9 KB   16/16 KB   204/204 KB   599/2891 KB   9/9 KB   16/16 KB   204/204 KB   603/2891 KB   9/9 KB   16/16 KB   204/204 KB   607/2891 KB   9/9 KB   16/16 KB   204/204 KB   611/2891 KB   9/9 KB   16/16 KB   204/204 KB   615/2891 KB   9/9 KB   16/16 KB   204/204 KB   619/2891 KB   9/9 KB   16/16 KB   204/204 KB   623/2891 KB   9/9 KB   16/16 KB   204/204 KB   627/2891 KB   9/9 KB   16/16 KB   204/204 KB   631/2891 KB   9/9 KB   16/16 KB   204/204 KB   635/2891 KB   9/9 KB   16/16 KB   204/204 KB   639/2891 KB   9/9 KB   16/16 KB   204/204 KB   643/2891 KB   9/9 KB   16/16 KB   204/204 KB   647/2891 KB   9/9 KB   16/16 KB   204/204 KB   651/2891 KB   9/9 KB   16/16 KB   204/204 KB   655/2891 KB   9/9 KB   16/16 KB   204/204 KB   659/2891 KB   9/9 KB   16/16 KB   204/204 KB   663/2891 KB   9/9 KB   16/16 KB   204/204 KB   667/2891 KB   9/9 KB   16/16 KB   204/204 KB   671/2891 KB   9/9 KB   16/16 KB   204/204 KB   675/2891 KB   9/9 KB   16/16 KB   204/204 KB   679/2891 KB   9/9 KB   16/16 KB   204/204 KB   683/2891 KB                                                                                                 Downloaded: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.jar (16 KB at 109.1 KB/sec)
#>                                                Downloaded: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.jar (9 KB at 60.0 KB/sec)
#> Downloaded: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.jar (204 KB at 1416.3 KB/sec)
#> 687/2891 KB                                    691/2891 KB   695/2891 KB   699/2891 KB   703/2891 KB   707/2891 KB   711/2891 KB   715/2891 KB   719/2891 KB   723/2891 KB   727/2891 KB   731/2891 KB   735/2891 KB   739/2891 KB   743/2891 KB   747/2891 KB   751/2891 KB   755/2891 KB   759/2891 KB   763/2891 KB   767/2891 KB   771/2891 KB   775/2891 KB   779/2891 KB   783/2891 KB   787/2891 KB   791/2891 KB   795/2891 KB   799/2891 KB   803/2891 KB   807/2891 KB   811/2891 KB   815/2891 KB   819/2891 KB   823/2891 KB   827/2891 KB   831/2891 KB   835/2891 KB   839/2891 KB   843/2891 KB   847/2891 KB   851/2891 KB   855/2891 KB   859/2891 KB   863/2891 KB   867/2891 KB   871/2891 KB   875/2891 KB   879/2891 KB   883/2891 KB   887/2891 KB   891/2891 KB   895/2891 KB   899/2891 KB   903/2891 KB   907/2891 KB   911/2891 KB   915/2891 KB   919/2891 KB   923/2891 KB   927/2891 KB   931/2891 KB   935/2891 KB   939/2891 KB   943/2891 KB   947/2891 KB   951/2891 KB   955/2891 KB   959/2891 KB   963/2891 KB   967/2891 KB   971/2891 KB   975/2891 KB   979/2891 KB   983/2891 KB   987/2891 KB   991/2891 KB   995/2891 KB   999/2891 KB   1003/2891 KB   1007/2891 KB   1011/2891 KB   1015/2891 KB   1019/2891 KB   1023/2891 KB   1027/2891 KB   1031/2891 KB   1035/2891 KB   1039/2891 KB   1043/2891 KB   1047/2891 KB   1051/2891 KB   1055/2891 KB   1059/2891 KB   1063/2891 KB   1067/2891 KB   1071/2891 KB   1075/2891 KB   1079/2891 KB   1083/2891 KB   1087/2891 KB   1091/2891 KB   1095/2891 KB   1099/2891 KB   1103/2891 KB   1107/2891 KB   1111/2891 KB   1115/2891 KB   1119/2891 KB   1123/2891 KB   1127/2891 KB   1131/2891 KB   1135/2891 KB   1139/2891 KB   1142/2891 KB   1146/2891 KB   1150/2891 KB   1154/2891 KB   1158/2891 KB   1162/2891 KB   1166/2891 KB   1170/2891 KB   1174/2891 KB   1178/2891 KB   1182/2891 KB   1186/2891 KB   1190/2891 KB   1194/2891 KB   1198/2891 KB   1202/2891 KB   1206/2891 KB   1210/2891 KB   1214/2891 KB   1218/2891 KB   1222/2891 KB   1226/2891 KB   1230/2891 KB   1234/2891 KB   1238/2891 KB   1242/2891 KB   1246/2891 KB   1250/2891 KB   1254/2891 KB   1258/2891 KB   1262/2891 KB   1266/2891 KB   1270/2891 KB   1274/2891 KB   1278/2891 KB   1282/2891 KB   1286/2891 KB   1290/2891 KB   1294/2891 KB   1298/2891 KB   1302/2891 KB   1306/2891 KB   1310/2891 KB   1314/2891 KB   1318/2891 KB   1322/2891 KB   1326/2891 KB   1330/2891 KB   1334/2891 KB   1338/2891 KB   1342/2891 KB   1346/2891 KB   1350/2891 KB   1354/2891 KB   1358/2891 KB   1362/2891 KB   1366/2891 KB   1370/2891 KB   1374/2891 KB   1378/2891 KB   1382/2891 KB   1386/2891 KB   1390/2891 KB   1394/2891 KB   1398/2891 KB   1402/2891 KB   1406/2891 KB   1410/2891 KB   1414/2891 KB   1418/2891 KB   1422/2891 KB   1426/2891 KB   1430/2891 KB   1434/2891 KB   1438/2891 KB   1442/2891 KB   1446/2891 KB   1450/2891 KB   1454/2891 KB   1458/2891 KB   1462/2891 KB   1466/2891 KB   1470/2891 KB   1474/2891 KB   1478/2891 KB   1482/2891 KB   1486/2891 KB   1490/2891 KB   1494/2891 KB   1498/2891 KB   1502/2891 KB   1506/2891 KB   1510/2891 KB   1514/2891 KB   1518/2891 KB   1522/2891 KB   1526/2891 KB   1530/2891 KB   1534/2891 KB   1538/2891 KB   1542/2891 KB   1546/2891 KB   1550/2891 KB   1554/2891 KB   1558/2891 KB   1562/2891 KB   1566/2891 KB   1570/2891 KB   1574/2891 KB   1578/2891 KB   1582/2891 KB   1586/2891 KB   1590/2891 KB   1594/2891 KB   1598/2891 KB   1602/2891 KB   1606/2891 KB   1610/2891 KB   1614/2891 KB   1618/2891 KB   1622/2891 KB   1626/2891 KB   1630/2891 KB   1634/2891 KB   1638/2891 KB   1642/2891 KB   1646/2891 KB   1650/2891 KB   1654/2891 KB   1658/2891 KB   1662/2891 KB   1666/2891 KB   1670/2891 KB   1674/2891 KB   1678/2891 KB   1682/2891 KB   1686/2891 KB   1690/2891 KB   1694/2891 KB   1698/2891 KB   1702/2891 KB   1706/2891 KB   1710/2891 KB   1714/2891 KB   1718/2891 KB   1722/2891 KB   1726/2891 KB   1730/2891 KB   1734/2891 KB   1738/2891 KB   1742/2891 KB   1746/2891 KB   1750/2891 KB   1754/2891 KB   1758/2891 KB   1762/2891 KB   1766/2891 KB   1770/2891 KB   1774/2891 KB   1778/2891 KB   1782/2891 KB   1786/2891 KB   1790/2891 KB   1794/2891 KB   1798/2891 KB   1802/2891 KB   1806/2891 KB   1810/2891 KB   1814/2891 KB   1818/2891 KB   1821/2891 KB   1825/2891 KB   1829/2891 KB   1833/2891 KB   1837/2891 KB   1841/2891 KB   1845/2891 KB   1849/2891 KB   1853/2891 KB   1857/2891 KB   1861/2891 KB   1865/2891 KB   1869/2891 KB   1873/2891 KB   1877/2891 KB   1881/2891 KB   1885/2891 KB   1889/2891 KB   1893/2891 KB   1897/2891 KB   1901/2891 KB   1905/2891 KB   1909/2891 KB   1913/2891 KB   1917/2891 KB   1921/2891 KB   1925/2891 KB   1929/2891 KB   1933/2891 KB   1937/2891 KB   1941/2891 KB   1945/2891 KB   1949/2891 KB   1953/2891 KB   1957/2891 KB   1961/2891 KB   1965/2891 KB   1969/2891 KB   1973/2891 KB   1977/2891 KB   1981/2891 KB   1985/2891 KB   1989/2891 KB   1993/2891 KB   1997/2891 KB   2001/2891 KB   2005/2891 KB   2009/2891 KB   2013/2891 KB   2017/2891 KB   2021/2891 KB   2025/2891 KB   2029/2891 KB   2033/2891 KB   2037/2891 KB   2041/2891 KB   2045/2891 KB   2049/2891 KB   2053/2891 KB   2057/2891 KB   2061/2891 KB   2065/2891 KB   2069/2891 KB   2073/2891 KB   2077/2891 KB   2081/2891 KB   2085/2891 KB   2089/2891 KB   2093/2891 KB   2097/2891 KB   2101/2891 KB   2105/2891 KB   2109/2891 KB   2113/2891 KB   2117/2891 KB   2121/2891 KB   2125/2891 KB   2129/2891 KB   2133/2891 KB   2137/2891 KB   2141/2891 KB   2145/2891 KB   2149/2891 KB   2153/2891 KB   2157/2891 KB   2161/2891 KB   2165/2891 KB   2169/2891 KB   2173/2891 KB   2177/2891 KB   2181/2891 KB   2185/2891 KB   2189/2891 KB   2193/2891 KB   2197/2891 KB   2201/2891 KB   2205/2891 KB   2209/2891 KB   2213/2891 KB   2217/2891 KB   2221/2891 KB   2225/2891 KB   2229/2891 KB   2233/2891 KB   2237/2891 KB   2241/2891 KB   2245/2891 KB   2249/2891 KB   2253/2891 KB   2257/2891 KB   2261/2891 KB   2265/2891 KB   2269/2891 KB   2273/2891 KB   2277/2891 KB   2281/2891 KB   2285/2891 KB   2289/2891 KB   2293/2891 KB   2297/2891 KB   2301/2891 KB   2305/2891 KB   2309/2891 KB   2313/2891 KB   2317/2891 KB   2321/2891 KB   2325/2891 KB   2329/2891 KB   2333/2891 KB   2337/2891 KB   2341/2891 KB   2345/2891 KB   2349/2891 KB   2353/2891 KB   2357/2891 KB   2361/2891 KB   2365/2891 KB   2369/2891 KB   2373/2891 KB   2377/2891 KB   2381/2891 KB   2385/2891 KB   2389/2891 KB   2392/2891 KB   2396/2891 KB   2400/2891 KB   2404/2891 KB   2408/2891 KB   2412/2891 KB   2416/2891 KB   2420/2891 KB   2424/2891 KB   2428/2891 KB   2432/2891 KB   2436/2891 KB   2440/2891 KB   2444/2891 KB   2448/2891 KB   2452/2891 KB   2456/2891 KB   2460/2891 KB   2464/2891 KB   2468/2891 KB   2472/2891 KB   2476/2891 KB   2480/2891 KB   2484/2891 KB   2488/2891 KB   2492/2891 KB   2496/2891 KB   2500/2891 KB   2504/2891 KB   2508/2891 KB   2512/2891 KB   2516/2891 KB   2520/2891 KB   2524/2891 KB   2528/2891 KB   2532/2891 KB   2536/2891 KB   2540/2891 KB   2544/2891 KB   2548/2891 KB   2552/2891 KB   2556/2891 KB   2560/2891 KB   2564/2891 KB   2568/2891 KB   2572/2891 KB   2576/2891 KB   2580/2891 KB   2584/2891 KB   2588/2891 KB   2592/2891 KB   2596/2891 KB   2600/2891 KB   2604/2891 KB   2608/2891 KB   2612/2891 KB   2616/2891 KB   2620/2891 KB   2624/2891 KB   2628/2891 KB   2632/2891 KB   2636/2891 KB   2640/2891 KB   2644/2891 KB   2648/2891 KB   2652/2891 KB   2656/2891 KB   2660/2891 KB   2664/2891 KB   2668/2891 KB   2672/2891 KB   2676/2891 KB   2680/2891 KB   2684/2891 KB   2688/2891 KB   2692/2891 KB   2696/2891 KB   2700/2891 KB   2704/2891 KB   2708/2891 KB   2712/2891 KB   2716/2891 KB   2720/2891 KB   2724/2891 KB   2728/2891 KB   2732/2891 KB   2736/2891 KB   2740/2891 KB   2744/2891 KB   2748/2891 KB   2752/2891 KB   2756/2891 KB   2760/2891 KB   2764/2891 KB   2768/2891 KB   2772/2891 KB   2776/2891 KB   2780/2891 KB   2784/2891 KB   2788/2891 KB   2792/2891 KB   2796/2891 KB   2800/2891 KB   2804/2891 KB   2808/2891 KB   2812/2891 KB   2816/2891 KB   2820/2891 KB   2824/2891 KB   2828/2891 KB   2832/2891 KB   2836/2891 KB   2840/2891 KB   2844/2891 KB   2848/2891 KB   2852/2891 KB   2856/2891 KB   2860/2891 KB   2864/2891 KB   2868/2891 KB   2872/2891 KB   2876/2891 KB   2880/2891 KB   2884/2891 KB   2888/2891 KB   2891/2891 KB                  Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar (2891 KB at 10989.0 KB/sec)
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD SUCCESS
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time: 2.274 s
#> [INFO] Finished at: 2022-08-22T15:38:17+00:00
#> [INFO] Final Memory: 15M/57M
#> [INFO] ------------------------------------------------------------------------
#> /home/runner/.m2/repository/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar
fetch_artifact(coordinates = as.coordinates("org.junit.jupiter","junit-jupiter-api","5.9.0"))
#> /home/runner/.m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.0/junit-jupiter-api-5.9.0.jar