Fetch an artifact from a repository into the local .m2 cache

Fetch an artifact from a repository into the local .m2 cache

fetch_artifact(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  ...,
  coordinates = NULL,
  artifact = NULL,
  repoUrl = getOption("rmaven.default_repos"),
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug")
)

Arguments

groupId

optional, the maven groupId,

artifactId

optional, the maven artifactId,

version

optional, the maven version,

...

other maven coordinates such as classifier or packaging

coordinates

optional, but if not supplied groupId and artifactId must be, coordinates as a coordinates object (see as.coordinates())

artifact

optional, coordinates as an artifact string groupId:artifactId:version[:packaging[:classifier]] string

repoUrl

the URLs of the repositories to check (defaults to maven central)

nocache

normally artifacts are only fetched if required, nocache forces fetching

verbose

how much output from maven, one of "normal", "quiet", "debug"

Value

the .m2 path of the artifact

Examples

# \donttest{
# This code can take quite a while to run as has to
# download a lot of plugins, especially on first run
fetch_artifact(artifact="com.google.guava:guava:31.1-jre")
#> executing: /home/runner/.cache/rmaven/mvnw org.apache.maven.plugins:maven-dependency-plugin:3.3.0:get -Dartifact=com.google.guava:guava:31.1-jre -B -s '/tmp/Rtmps1mjsA/rmaven/settings.xml'
#> [INFO] Scanning for projects...
#> [INFO]                                                                         
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Building Maven Stub Project (No POM) 1
#> [INFO] ------------------------------------------------------------------------
#> [INFO] 
#> [INFO] --- maven-dependency-plugin:3.3.0:get (default-cli) @ standalone-pom ---
#> [INFO] Resolving com.google.guava:guava:jar:31.1-jre with transitive dependencies
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.pom (11 KB at 45.8 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/31.1-jre/guava-parent-31.1-jre.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/31.1-jre/guava-parent-31.1-jre.pom (15 KB at 757.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.pom (3 KB at 112.2 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/26.0-android/guava-parent-26.0-android.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava-parent/26.0-android/guava-parent-26.0-android.pom (10 KB at 473.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.pom (3 KB at 139.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.pom (5 KB at 174.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.pom (3 KB at 107.5 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.pom (3 KB at 96.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_parent/2.11.0/error_prone_parent-2.11.0.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_parent/2.11.0/error_prone_parent-2.11.0.pom (11 KB at 582.4 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.pom
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.pom (3 KB at 142.0 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.jar
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar (5 KB at 66.3 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.jar (20 KB at 278.1 KB/sec)
#> [INFO] Downloading: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.jar
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/listenablefuture/9999.0-empty-to-avoid-conflict-with-guava/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar (3 KB at 25.9 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/errorprone/error_prone_annotations/2.11.0/error_prone_annotations-2.11.0.jar (16 KB at 173.4 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/j2objc/j2objc-annotations/1.3/j2objc-annotations-1.3.jar (9 KB at 94.2 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/org/checkerframework/checker-qual/3.12.0/checker-qual-3.12.0.jar (204 KB at 1359.6 KB/sec)
#> [INFO] Downloaded: https://repo.maven.apache.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar (2891 KB at 12194.6 KB/sec)
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD SUCCESS
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time: 1.995 s
#> [INFO] Finished at: 2022-09-15T19:38:35+00:00
#> [INFO] Final Memory: 15M/54M
#> [INFO] ------------------------------------------------------------------------
#> /home/runner/.cache/rmaven/.m2/repository/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar
fetch_artifact(coordinates = as.coordinates("org.junit.jupiter","junit-jupiter-api","5.9.0"))
#> /home/runner/.cache/rmaven/.m2/repository/org/junit/jupiter/junit-jupiter-api/5.9.0/junit-jupiter-api-5.9.0.jar
# }